move beam state obj to new BeamState class
